Kafka

+++ ProductService/src/main/java/com/wolfhack/cloud/product/repository/RamRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.mongodb.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RamRepository extends MongoRepository<Ram, Double> {
+public interface RamRepository extends MongoRepository<Ram, Long> {
 
     @Query("{ $text: { $search: ?0 }}")
     Page<Ram> searchCpusByQuery(String query, Pageable pageable);
