+++ ProductService/src/main/java/com/wolfhack/cloud/product/repository/SsdRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.mongodb.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SsdRepository extends MongoRepository<Ssd, Double> {
+public interface SsdRepository extends MongoRepository<Ssd, Long> {
 
     @Query("{ $text: { $search: ?0 }}")
     Page<Ssd> searchCpusByQuery(String query, Pageable pageable);
